python-server/requirements.txt
numpy==2.2.6
opencv-python==4.12.0.88
paho-mqtt==1.6.1
